Removed type warnings.

+++ I Bike CPH/CykelPlanen/Classes/SMArrivalInformation.h
@@ -6,20 +6,20 @@
 //  Copyright (c) 2013 City of Copenhagen. All rights reserved.
 //
 
-#import <Foundation/Foundation.h>
-#import "SMStationInfo.h"
 #import "SMArrivalMapping.h"
+#import "SMStationInfo.h"
+#import <Foundation/Foundation.h>
 
 /**
- * TODO: Arrival information for station. 
+ * TODO: Arrival information for station.
  */
 @interface SMArrivalInformation : NSObject
 
-@property(nonatomic, strong) SMStationInfo* station;
-@property(nonatomic, strong) NSMutableArray* mappings;
+@property(nonatomic, strong) SMStationInfo *station;
+@property(nonatomic, strong) NSMutableArray *mappings;
 
--(void)addArrivalTime:(SMTime*)pTime forDays:(NSArray*)days;
--(void)addDepartureTime:(SMTime*)pTime forDays:(NSArray*)days;
--(SMArrivalMapping*)mappingForDayAtIndex:(int)index;
--(BOOL)hasInfoForDayAtIndex:(int)index;
+- (void)addArrivalTime:(SMTime *)pTime forDays:(NSArray *)days;
+- (void)addDepartureTime:(SMTime *)pTime forDays:(NSArray *)days;
+- (SMArrivalMapping *)mappingForDayAtIndex:(NSInteger)index;
+- (BOOL)hasInfoForDayAtIndex:(NSUInteger)index;
 @end
